Move formatting code from model to util
